Change port and enforce 16 characters user password length

+++ renew_podman_secrets.sh
@@ -2,10 +2,17 @@
 
 umask 077
 
-read -r -s -p "Enter proxy user password: " user_password
+read -r -s -p "Enter proxy user password (16 characters): " user_password
 echo
 
-printf '%s' "$user_password" > mtproxy_user_pass
+password_length=${#user_password}
+
+if [ ! ${password_length} -eq 16 ]; then
+    echo "User password length must be exactly 16 characters of length"
+    exit 1
+fi
+
+od -An -v -tx1 "$user_password" | tr -d ' \n' > mtproxy_user_pass
 
 podman secret rm mtproxy-user-pass 2>/dev/null || true
 podman secret create mtproxy-user-pass mtproxy_user_pass
